Sun Spectrum and loop button added
Visuals for loop will be improved.

The Sun Spectrum is my pride so far. It combines the usage of
setnumbands that I worked so hard to understand due to no documentation,
and a lot of work on figuring out how to rotate lines correctly.
It also uses my research on the equalizer bands.